Media queries and fav-icon
Here i have added the media quieries for the site and i have made it responsive for all screen sizes down to the iphone SE example given on the google inspect tool. as well as adding a mobile nav so when the viewport gets to a certain size it will convert the nav. I have also implemented a fav-icon for my portfolio that i designed in photoshop to give it a bit of life.