EDIT # 2 - Bug Fix
Found a bug where the coords for the sections where not being calculated correctly on an actual mobile device not desktop. I believe this is caused by the calculation removing the navHeight from the top coords in order to leave room for that in fullscreen when it needs to remain stuck to the top of the screen. To solve this i have implemented a simple check to see if the mobile layout is active or the desktop one. and based on that decide whether to remove the navHeight from the coords or not.