Added projects and updated text
I have added 2 new projects to the portfolio, the first is a real-time chat application with firebase integration. It allows users to speak to other users in real-time. Users can add friends, send text or image messages and view all chat history between them and another user.

The second project is a remake of an older project i made for a client. I made them a gallery website around a year ago, and as i have progressed considerably in development i felt it would be good to get a new site for the client that is much better. I liaised directly with the client to come up with a design and color palette that the client was happy with. The i got to work making this site from scratch using React, Vite, Firebase, and some third party libraries. To date this is my proudest complete site. I learned much more about using firebase and its many features such as cloud storage, firebase authorization and firebase firestore. This amount of exposure to firebase tools has really helped to improve my workflow when talking to firebase and integrating it into my applications.